rename spectra
